+++ packages/better_player_web/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Added: Support for `drmSecurityLevel` mapping to Shaka Player's `videoRobustness`.
+
 ## 1.0.1
 
 * Added example application for web platform.